updated task 16

+++ 16-binary_tree_is_perfect.c
@@ -8,7 +8,7 @@
 
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	size_t checker = 0;
+	int checker = 0;
 
 	if (tree == NULL)
 		return (0);
